% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_afni_cmd.R
\name{build_afni_cmd}
\alias{build_afni_cmd}
\title{build_afni_cmd}
\usage{
build_afni_cmd(afni_list)
}
\arguments{
\item{afni_list}{list that contains the program name and options to create command}
}
\value{
return afni command
}
\description{
build_afni_cmd
}
\examples{
build_afni_cmd(afni_list = list(program = "3dcalc",
                                opt = list(a = "a.nii.gz",
                                           b = "b.nii.gz",
                                           expr = "'(a*b)'"),
                                           prefix = c.nii.gz))
}
\concept{support}
