% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_wrangling_report.R
\name{create_data_wrangling_report}
\alias{create_data_wrangling_report}
\title{create_data_wrangling_report}
\usage{
create_data_wrangling_report(
  data_wrangling,
  out_dir = NULL,
  out_prefix = NULL,
  out_suffix = NULL,
  quiet = TRUE
)
}
\arguments{
\item{data_wrangling}{data wrangling output}

\item{out_dir}{output directory (default: NULL) if null, saves to current working directory}

\item{out_prefix}{output file prefix (default: NULL)}

\item{out_suffix}{output file suffix (default: NULL)}

\item{quiet}{suppress message from rmarkdown::render() (default: TRUE)}
}
\value{

}
\description{
create_data_wrangling_report
}
