% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrix.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{create_design_matrix}
\usage{
create_design_matrix(psy_var, phys_var, ppi_var, nuisance_var = NULL)
}
\arguments{
\item{psy_var}{data of psy variable}

\item{phys_var}{data of phys variable}

\item{ppi_var}{data of ppi variable}

\item{nuisance_var}{data of nuisance variables}
}
\value{
design matrix
}
\description{
create_design_matrix
}
\examples{
# to be added
}
