% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hrf_afni.R
\name{create_hrf_afni}
\alias{create_hrf_afni}
\title{create_hrf_afni}
\usage{
create_hrf_afni(hrf, tr, upsample_factor = NULL)
}
\arguments{
\item{hrf}{name of the hemodynamic response function. can be gam, block or spmg1}

\item{tr}{retrieval time (tr) in seconds}

\item{upsample_factor}{number of volumes or time points}
}
\description{
create_hrf_afni
}
\examples{
hrf <- create_hrf_afni("spmg1", 1.5, 16)
head(hrf)
}
