% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ppi_var.R
\name{create_ppi_var}
\alias{create_ppi_var}
\title{create_ppi_var}
\usage{
create_ppi_var(
  psy_var,
  phys_var,
  hrf,
  tr,
  n_volumes,
  upsample_factor = NULL,
  deconvolve = TRUE,
  afni_path = NULL
)
}
\arguments{
\item{psy_var}{psychological variable data (e.g., psy var that is contrast-coded and upsampled)}

\item{phys_var}{physiological variable data (e.g., psy var that is detrended and upsampled)}

\item{hrf}{hemodynamic response function (hrf) time series data}

\item{tr}{repetition time (tr) in seconds}

\item{n_volumes}{number of volumes or time points}

\item{upsample_factor}{factor to upsample data by for convolution and deconvolution step (default: NULL)}

\item{deconvolve}{option to perform or not perform deconvolution step (default: TRUE)}

\item{afni_path}{path to afni directory (default: NULL)}
}
\value{
a list of datasets for each data wrangling step of the ppi variables
}
\description{
creates data for each step of the ppi variable by wrapping the functions: \code{convolve_afni()} and \code{downsample()}
}
\concept{data_wrangling_wrapper_functions}
