% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_psy_var.R
\name{create_psy_var}
\alias{create_psy_var}
\title{create_psy_var}
\usage{
create_psy_var(
  events,
  contrast_table,
  hrf,
  tr,
  n_volumes,
  upsample_factor = NULL
)
}
\arguments{
\item{events}{events data with columns of onset, duration, and trial_type}

\item{contrast_table}{table of contrast codes}

\item{hrf}{hemodynamic response function (hrf) time series}

\item{tr}{retrieval time (tr) in seconds}

\item{n_volumes}{number of volumes or time points}

\item{upsample_factor}{factor to upsample the trial_type_by_volume data during the convolution step}
}
\value{
a list of datasets for each data wrangling step of the psychological variables
}
\description{
create_psy_var
}
\examples{
# to be added
}
