% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_series_atlas_figure.R
\name{create_time_series_atlas_figure}
\alias{create_time_series_atlas_figure}
\title{create_time_series_atlas_figure}
\usage{
create_time_series_atlas_figure(data, title = NULL, nrow = NULL,
  ncol = NULL, scales = NULL)
}
\arguments{
\item{data}{data to create time series figure}

\item{title}{title of figure (default: NULL)}
}
\value{
time series figure with volume on the x-axis and value on the y-axis
}
\description{
create_time_series_atlas_figure
}
\concept{visualization}
