% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_series_heatmap_figure.R
\name{create_time_series_heatmap_figure}
\alias{create_time_series_heatmap_figure}
\title{create_time_series_heatmap_figure}
\usage{
create_time_series_heatmap_figure(data, title = NULL,
  transpose = FALSE)
}
\arguments{
\item{data}{data of volume/time points (row) by region of interest (column) for heatmap figure}

\item{title}{title of figure (default: NULL)}
}
\value{
heatmap of volume by region of interest
}
\description{
create_time_series_heatmap_figure
}
\concept{visualization}
