% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_fig.R
\name{create_ts_fig}
\alias{create_ts_fig}
\title{create_ts_fig}
\usage{
create_ts_fig(data, title = NULL, scales = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{data}{data to create time series figure}

\item{title}{title of figure (default: NULL)}
}
\value{
time series figure with volume on the x-axis and value on the y-axis
}
\description{
create_ts_fig
}
\concept{visualization}
