% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{data_wrangling}
\alias{data_wrangling}
\title{data_wrangling}
\usage{
data_wrangling(psy_events_data, psy_contrast_table, phys_data,
  detrend_factor, hrf_name, tr, n_volumes, upsample_factor = NULL,
  deconvolve = TRUE, nuisance_var = NULL)
}
\arguments{
\item{psy_events_data}{events data with columns of onset, duration, and trial_type}

\item{psy_contrast_table}{contrast code}

\item{phys_data}{time series data of physiological variable (seed ROI)}

\item{detrend_factor}{factor to detrend time series}

\item{hrf_name}{name of hemodynamic response function (hrf) to use during convolution and deconvolution step}

\item{tr}{repition time (tr) in seconds}

\item{n_volumes}{number of volumes or time points}

\item{upsample_factor}{factor to upsample psy and phys data for convolution and deconvolution step}

\item{deconvolve}{perform deconvolution of phys and convolve of ppi term (also known as deconvolve-convolve or reconvolve step)}

\item{nuisance_var}{data of nuisance variables to include in design matrix}
}
\description{
uber data wrangling wrapper function to create data for each step of data wrangling for each psy, phys, and ppi variables to ultimately create the design matrix. This function is a wrapper of \code{create_hrf_afni()}, \code{create_psy_var()}, \code{create_phys_var()}, \code{create_ppi_var()}, and \code{create_design_matrix()}
}
\examples{
# to be added
}
\concept{data_wrangling_uber_functions}
