% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_afni.R
\name{deconvolve_afni}
\alias{deconvolve_afni}
\title{deconvolve_afni}
\usage{
deconvolve_afni(data, hrf, afni_path = NULL)
}
\arguments{
\item{data}{data to deconvolve}

\item{hrf}{hemodynamic response function (hrf) to deconvolve data}

\item{afni_path}{path to afni directory (default: NULL)}
}
\description{
deconvolve_afni
}
\concept{data_wrangling}
