% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{detrend}
\usage{
detrend(data, degree)
}
\arguments{
\item{data}{data to be detrended}

\item{degree}{degree for data to be detrended}
}
\value{
detrended data
}
\description{
detrend
}
\concept{data_wrangling}
