% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{downsample}
\usage{
downsample(data, downsample_factor)
}
\arguments{
\item{data}{data to be upsampled}

\item{downsample_factor}{factor to downsample data}
}
\description{
downsample
}
\examples{
# to be added
}
\concept{data_wrangling}
