% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_afni_cmd.R
\name{execute_afni_cmd}
\alias{execute_afni_cmd}
\title{execute_afni_cmd}
\usage{
execute_afni_cmd(afni_cmd, afni_path = NULL)
}
\arguments{
\item{afni_cmd}{afni command to execute within R}

\item{afni_path}{afni path (default: NULL). the default option will use \code{afnir::get_afni()} to obtain the path}
}
\description{
execute_afni_cmd
}
\examples{
to be added
}
