% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_time_series_mean.R
\name{extract_time_series_mean}
\alias{extract_time_series_mean}
\title{extract_time_series_mean}
\usage{
extract_time_series_mean(data, mask, labels = NULL)
}
\arguments{
\item{data}{data to extract mean time series from}

\item{mask}{mask to apply and extract mean time series from. mask can also be a 4D array with a different regions of interest in each 4th dimension}

\item{labels}{labels the correspond to the mask regions of interest (default: NULL)}
}
\value{
data.frame of time/volume (row) by region
}
\description{
extract_time_series_mean
}
\concept{data_extraction}
