% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_time_series_mean_rw.R
\name{extract_time_series_mean_rw}
\alias{extract_time_series_mean_rw}
\title{extract_time_series_mean_rw}
\usage{
extract_time_series_mean_rw(in_path_data, in_path_mask,
  in_path_labels = NULL, out_path, overwrite = FALSE)
}
\arguments{
\item{in_path_data}{path to nifti data for time series extraction}

\item{in_path_mask}{path to nifti mask applied to data for time series extraction (mask can be a 4D nifti file with a different region of interest in each 4th dimension)}

\item{in_path_labels}{path to labels .csv file with a single column of labels with no header (default: NULL)}

\item{out_path}{path to output .csv file of time/volume (rows) by regions of interest (column)}

\item{overwrite}{option to overwrite out_path file (default: FALSE)}
}
\description{
extract_time_series_mean_rw
}
\concept{data_extraction}
