% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_time_series_rw_fsl.R
\name{extract_time_series_rw_fsl}
\alias{extract_time_series_rw_fsl}
\title{extract_time_series_rw_fsl}
\usage{
extract_time_series_rw_fsl(in_nii, in_mask_nii, out_path,
  statistic = "mean", overwrite = F, fsl_path = NULL)
}
\arguments{
\item{fsl_path}{}
}
\description{
extract_time_series_rw_fsl
}
\concept{data_extraction}
