% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO_url.R
\name{getGEO_url}
\alias{getGEO_url}
\title{Replace getGEO function}
\usage{
getGEO_url(GSE)
}
\arguments{
\item{GSE}{GSE number of GEO database}
}
\value{
GSE dataset download address
}
\description{
Replace getGEO function
}
\keyword{Replace}
\keyword{function}
\keyword{getGEO}
