% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_glm_roi2roi.R
\name{model_glm_roi2roi}
\alias{model_glm_roi2roi}
\title{model_glm_roi2roi}
\usage{
model_glm_roi2roi(target_roi, design_matrix)
}
\arguments{
\item{target_roi}{data to predict. can be also be a list or multiple columns if there are more than one target roi}

\item{design_matrix}{predictors. can be also be list or a 3d array if there are more than one set of predictors}
}
\description{
model_glm_roi2roi
}
\concept{analysis}
