% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI.R
\docType{package}
\name{PPI}
\alias{PPI}
\title{Protein-Protein Interaction
This package is used to construct protein-protein interaction network, and build-in 4 algorithms to calculate Hub gene and provide gene ID conversion}
