% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_data.R
\name{PPI_data}
\alias{PPI_data}
\title{Building PPI network data}
\usage{
PPI_data(name)
}
\arguments{
\item{name}{Set of genes named ENTREZID}
}
\value{
PPI network data
}
\description{
Building PPI network data
}
\keyword{Protein}
\keyword{interaction}
