% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_hub.R
\name{PPI_hub}
\alias{PPI_hub}
\title{Using algorithm to calculate Hub gene}
\usage{
PPI_hub(PPI_data_symbol)
}
\arguments{
\item{PPI_data_symbol}{A set of interaction data}
}
\value{
data_mysuanfa (The result of hub gene found by the algorithm)
}
\description{
Using algorithm to calculate Hub gene
}
\keyword{Hub}
\keyword{PPI(Protein}
\keyword{Protein}
\keyword{gene)}
\keyword{interaction}
