% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list_data.R
\name{save_list_data}
\alias{save_list_data}
\title{save_list_data}
\usage{
save_list_data(data, out_dir, out_prefix = NULL, out_suffix = NULL)
}
\arguments{
\item{data}{list data to save}

\item{out_dir}{directory to save list data}

\item{out_prefix}{prefix of files to save (default: NULL)}

\item{out_suffix}{suffix of the files to save (default: NULL)}
}
\value{

}
\description{
save_list_data
}
\concept{helper}
