% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list_data_fig.R
\name{save_list_data_fig}
\alias{save_list_data_fig}
\title{save_list_data_fig}
\usage{
save_list_data_fig(data, in_file)
}
\arguments{
\item{in_file}{}
}
\value{

}
\description{
save_list_data_fig
}
