% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_min_max.R
\name{scale_min_max}
\alias{scale_min_max}
\title{scale_min_max}
\usage{
scale_min_max(data, min = NULL, max = NULL)
}
\arguments{
\item{data}{data to scale}

\item{min}{minimum value of the data (default: NULL)}

\item{max}{maximum value of the data (default: NULL)}
}
\description{
scale_min_max
}
