% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample.R
\name{upsample}
\alias{upsample}
\title{upsample}
\usage{
upsample(data, upsample_factor)
}
\arguments{
\item{data}{data to upsample}

\item{upsample_factor}{factor to upsample data}
}
\description{
upsample
}
\examples{
# to be added
}
\concept{data_wrangling}
