% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_time_series.R
\name{visualize_time_series}
\alias{visualize_time_series}
\title{visualize_time_series}
\usage{
visualize_time_series(
  data,
  title = NULL,
  scales = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{data}{data to create time series figure}

\item{title}{title of figure (default: NULL)}
}
\value{
time series figure with volume on the x-axis and value on the y-axis
}
\description{
visualize_time_series
}
\concept{visualization}
