% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_time_series_heatmap.R
\name{visualize_time_series_heatmap}
\alias{visualize_time_series_heatmap}
\title{visualize_time_series_heatmap}
\usage{
visualize_time_series_heatmap(data, scale_data = FALSE, title = NULL,
  variable_label = "variable", caption = NULL,
  numeric_variable_names = FALSE, palette = "RdBu", direction = 1,
  transpose = FALSE)
}
\arguments{
\item{data}{data of volume/time points (row) by variable (e.g., multiple regions of interest or predictors of design matrix) for heatmap figure}

\item{title}{title of figure (default: NULL)}
}
\value{
heatmap of volume by variable
}
\description{
visualize_time_series_heatmap
}
\concept{visualization}
