% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve_afni.R
\name{convolve_afni}
\alias{convolve_afni}
\title{convolve_afni}
\usage{
convolve_afni(data, hrf, tr, n_volumes, upsample_factor = NULL)
}
\arguments{
\item{data}{data to convolve}

\item{hrf}{hemodynamic response function (hrf) time series}

\item{tr}{repition time (tr) in seconds}

\item{n_volumes}{number of volumes or time points}

\item{upsample_factor}{factor to upsample data for convolution step}
}
\value{

}
\description{
convolve_afni
}
