% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_wrangling_report.R
\name{create_data_wrangling_report}
\alias{create_data_wrangling_report}
\title{create_data_wrangling_report}
\usage{
create_data_wrangling_report(
  data_wrangling,
  out_dir,
  out_prefix,
  out_suffix,
  quiet = T
)
}
\arguments{
\item{data_wrangling}{data wrangling output}

\item{out_dir}{output directory}

\item{out_prefix}{output file prefix}

\item{out_suffix}{output file suffix}

\item{quiet}{suppress message from rmarkdown::render()}
}
\value{

}
\description{
create_data_wrangling_report
}
