% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrix.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{create_design_matrix}
\usage{
create_design_matrix(psy_var, phys_var, ppi_var, nuisance_var = NULL)
}
\arguments{
\item{psy_var}{psy variable data to add to design matrix}

\item{phys_var}{phys variable data to add to design matrix}

\item{ppi_var}{ppi variable data to add to design matrix}

\item{nuisance_var}{nuisance variable data to add to design matrix}
}
\value{
design matrix
}
\description{
create_design_matrix
}
\examples{
# to be added
}
\concept{data_wrangling_wrapper_functions}
