% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phys_var.R
\name{create_phys_var}
\alias{create_phys_var}
\title{create_phys_var}
\usage{
create_phys_var(phys, detrend_factor, upsample_factor = NULL, hrf)
}
\arguments{
\item{phys}{time series data}

\item{detrend_factor}{factor to detrend the time series by}

\item{upsample_factor}{factor to upsample the time series by for the deconvolution step}

\item{hrf}{hemodynamic response function time series data}
}
\value{
a list of datasets for each data wrangling step of the physiological variables
}
\description{
create_phys_var
}
\examples{
# to be added
}
