% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phys_var.R
\name{create_phys_var}
\alias{create_phys_var}
\title{create_phys_var}
\usage{
create_phys_var(phys, detrend_factor, upsample_factor = NULL, hrf,
  afni_path = NULL, afni_quiet = FALSE)
}
\arguments{
\item{phys}{time series data}

\item{detrend_factor}{factor to detrend the time series}

\item{upsample_factor}{factor to upsample the time series for the deconvolution step (default: NULL)}

\item{hrf}{hemodynamic response function (hrf) time series data}

\item{afni_path}{path to afni directory (default: NULL)}
}
\value{
a list of datasets for each data wrangling step of the physiological variables
}
\description{
creates data for each wrangling step of the physiological variable using the data wrangling functions: \code{detrend()}, \code{upsample()}, and \code{deconvolve_afni()}
}
\examples{
df <- read_csv("examples/sub-control01_task-music_run-1_bold_space-subj_vox-32-24-38")
phys_ts_file <- "examples/sub-control01_task-music_run-1_bold_space-subj_vox-32-24-38.csv"
phys_ts <- read.csv(phys_ts_file, header = F)
hrf <- create_hrf_afni("spmg1", 3, NULL)
phys_var <- create_phys_var(phys_ts_file, 2, 16, hrf)
summary(test)
}
\concept{data_wrangling_wrapper_functions}
