% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{ppi}
\alias{ppi}
\title{Calculate the PPI (Plant Phenology Index)}
\usage{
ppi(dvi, zenith.angle, M = max(dvi) + 0.005, dvi.soil = 0.09, G = 0.5)
}
\arguments{
\item{dvi}{Difference vegetation index (DVI = NIR - red) using sun-sensor corrected
reflectances, for example from nadir-view MODIS NBAR}

\item{zenith.angle}{Solar zenith angles (in radians) corresponding to the \code{dvi} values.}

\item{M}{Canopy maximum of \code{dvi}, default is estimated
from given values plus small constant (0.005)}

\item{dvi.soil}{DVI of soil, default value is 0.09.}

\item{G}{Geometric function of leaf angular distribution.
Default value 0.5 is valid for flat and needle leaves,
when assuming a spherical leaf angle distribution.}
}
\value{
Values of PPI.
}
\description{
Function to calculate the PPI from sun-sensor corrected NIR (near-infrared) and red reflectances after Jin and Eklundh 2014
}
\details{
This function calculates the PPI, as given in Jin H, Eklundh L. 2014.
A physically based vegetation index for improved monitoring of plant
phenology. Remote Sensing of Environment 152: 512–525. DOI: 10.1016/j.rse.2014.07.010.

Default parameters are as in the paper.

No NA values are accepted (throws an error).
}
\examples{
# some MODIS NBAR data
data(modis)

# remove any NA
modis.na <- na.omit(modis)

# MODIS data has factor of 10000; zenith angle converted into radians
ppi.values <- ppi(dvi = (modis.na$nir - modis.na$red) / 10000,
                  zenith.angle = modis.na$solar.zenith.angle / 180 * pi)
plot(modis.na$date, ppi.values)

}
