% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_data.R
\name{PPI_data}
\alias{PPI_data}
\title{Building PPI network data}
\usage{
PPI_data(name)
}
\arguments{
\item{name}{Set of genes named ENTREZID (preferably in the format of interger)}
}
\value{
net (PPI network data)
}
\description{
Building PPI network data
}
\keyword{PPI(Protein}
\keyword{Protein}
\keyword{interaction)}
