% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_wrangling.R
\name{save_data_wrangling}
\alias{save_data_wrangling}
\title{save_data_wrangling}
\usage{
save_data_wrangling(data_wrangling, out_dir, out_prefix, out_suffix)
}
\arguments{
\item{data_wrangling}{data wrangling output}

\item{out_dir}{output directory}

\item{out_prefix}{output file prefix}

\item{out_suffix}{output file suffix}
}
\value{

}
\description{
save_data_wrangling
}
