% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_heatmap_figure_from_list_data.R
\name{save_heatmap_figure_from_list_data}
\alias{save_heatmap_figure_from_list_data}
\title{save_heatmap_figure_from_list_data}
\usage{
save_heatmap_figure_from_list_data(data, out_dir, out_prefix = NULL,
  out_suffix = NULL, transpose = FALSE, width = NA, height = NA,
  units = "in")
}
\arguments{
\item{data}{list data to create and save time series figure}

\item{out_dir}{directory to save figures}

\item{out_prefix}{prefix of figures to save (default: NULL)}

\item{out_suffix}{suffix of figures to save (default: NULL)}
}
\description{
save_heatmap_figure_from_list_data
}
\concept{helper}
