% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list_data_fig.R
\name{save_list_data_fig}
\alias{save_list_data_fig}
\title{save_ts_fig_from_list_data}
\usage{
save_list_data_fig(
  data,
  out_dir,
  out_prefix = NULL,
  out_suffix = NULL,
  scales = NULL,
  nrow = NULL,
  ncol = NULL,
  width = NA,
  height = NA,
  units = "in"
)
}
\arguments{
\item{data}{list data to create and save time series figure}

\item{out_dir}{directory to save figures}

\item{out_prefix}{prefix of figures to save (default: NULL)}

\item{out_suffix}{suffix of figures to save (default: NULL)}
}
\value{

}
\description{
save_ts_fig_from_list_data
}
\concept{helper}
